+++ R/fit-model.R
@@ -223,13 +223,14 @@ fitmod <- function(obj, ..., epp=FALSE, B0 = 1e5, B = 1e4, B.re = 3000, number_k
     eppd$hhsage <- eppd$sibmx <- NULL
 
   likdat <- prepare_likdat(eppd, fp)
+  fp$ancsitedata <- as.logical(length(likdat$anclik.dat$W.lst))
 
   if(fp$eppmod %in% c("rhybrid", "logrw", "rlogistic_rw")){  # THIS IS REALLY MESSY, NEED TO REFACTOR CODE
     fp$SIM_YEARS <- as.integer(likdat$lastdata.idx)
     fp$proj.steps <- seq(fp$ss$proj_start+0.5, fp$ss$proj_start-1+fp$SIM_YEARS+0.5, by=1/fp$ss$hiv_steps_per_year)
   } else
     fp$SIM_YEARS <- fp$ss$PROJ_YEARS
-  
+
 
   ## Prepare the EPP model
   tsEpidemicStart <- if(epp) fp$tsEpidemicStart else fp$ss$time_epi_start+0.5
@@ -245,6 +246,27 @@ fitmod <- function(obj, ..., epp=FALSE, B0 = 1e5, B = 1e4, B.re = 3000, number_k
     fp <- prepare_logrw(fp)
   else if(fp$eppmod == "rlogistic_rw")
     fp <- prepare_rlogistic_rw(fp)
+
+  ## Fit using optimization
+  if(optfit){
+    optfn <- function(theta, fp, likdat) lprior(theta, fp) + ll(theta, fp, likdat)
+    if(is.null(opt_init)){
+      X0 <- sample_prior(B0, fp)
+      lpost0 <- likelihood(X0, fp, likdat, log=TRUE) + prior(X0, fp, log=TRUE)
+      opt_init <- X0[which.max(lpost0)[1],]
+    }
+    opt <- optim(opt_init, optfn, fp=fp, likdat=likdat, method=opt_method, control=list(fnscale=-1, trace=4, maxit=1e3))
+    opt$fp <- fp
+    opt$likdat <- likdat
+    opt$param <- fnCreateParam(opt$par, fp)
+    opt$mod <- simmod(update(fp, list=opt$param))
+    if(epp)
+      class(opt) <- "eppopt"
+    else
+      class(opt) <- "specopt"
+    
+    return(opt)
+  }
 
   ## Fit using optimization
   if(optfit){